feat : 화상상담 입장 정보 조회 기능 구현 #S11P12C206-806

+++ backend/src/main/java/com/ssafy/kidslink/application/meetingschedule/repository/MeetingScheduleRepository.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface MeetingScheduleRepository extends JpaRepository<MeetingSchedule, Long> {
+public interface MeetingScheduleRepository extends JpaRepository<MeetingSchedule, Integer> {
     List<MeetingSchedule> findByTeacher(Teacher teahcer);
     List<MeetingSchedule> findByParent(Parent parent);
     List<MeetingSchedule> findByTeacherAndMeetingScheduleDate(Teacher teacher, LocalDate date);
